Update vocab init
